ConfigLoaders::Throttled (rename and minor tweaks)
